api/requirements.txt
fastapi
databases
sqlalchemy
httpx
asyncpg
pydantic
uvicorn
pydantic_settings
psycopg2-binary
python-dotenv